Refactor; remove unneeded functions
