+++ src/index.js
@@ -1,3 +1,4 @@
+import useBlocklyWorkspace from "./useBlocklyWorkspace";
 import BlocklyWorkspace from "./BlocklyWorkspace";
 
-export { BlocklyWorkspace };
+export { BlocklyWorkspace, useBlocklyWorkspace };
