+++ src/useBlocklyWorkspace.js
@@ -1,0 +1,110 @@
+import React from "react";
+import Blockly from "blockly";
+
+import debounce from "./debounce";
+
+function importFromXml(xml, workspace, onImportXmlError) {
+  try {
+    Blockly.Xml.domToWorkspace(Blockly.Xml.textToDom(xml), workspace);
+    return true;
+  } catch (e) {
+    if (onImportXmlError) {
+      onImportXmlError(e);
+    }
+    return false;
+  }
+}
+
+const useBlocklyWorkspace = ({
+  ref,
+  initialXml,
+  toolboxConfiguration,
+  workspaceConfiguration,
+  onWorkspaceChange,
+  onImportXmlError,
+}) => {
+  const [workspace, setWorkspace] = React.useState(null);
+  const [xml, setXml] = React.useState(initialXml);
+  const [didInitialImport, setDidInitialImport] = React.useState(false);
+
+  // we explicitly don't want to recreate the workspace when the configuration changes
+  // so, we'll keep it in a ref and update as necessary in an effect hook
+  const workspaceConfigurationRef = React.useRef(workspaceConfiguration);
+  React.useEffect(() => {
+    workspaceConfigurationRef.current = workspaceConfiguration;
+  }, [workspaceConfiguration]);
+
+  const toolboxConfigurationRef = React.useRef(toolboxConfiguration);
+  React.useEffect(() => {
+    toolboxConfigurationRef.current = toolboxConfiguration;
+    if (toolboxConfiguration && workspace) {
+      workspace.updateToolbox(toolboxConfiguration);
+    }
+  }, [toolboxConfiguration, workspace]);
+
+  const handleWorkspaceChanged = React.useCallback(
+    (newWorkspace) => {
+      if (onWorkspaceChange) {
+        onWorkspaceChange(newWorkspace);
+      }
+    },
+    [onWorkspaceChange]
+  );
+
+  // Initial mount
+  React.useEffect(() => {
+    const newWorkspace = Blockly.inject(ref.current, {
+      ...workspaceConfigurationRef.current,
+      toolbox: toolboxConfigurationRef.current,
+    });
+    setWorkspace(newWorkspace);
+    setDidInitialImport(false); // force a re-import if we recreate the workspace
+    handleWorkspaceChanged(newWorkspace);
+    newWorkspace.addChangeListener(() => {
+      handleWorkspaceChanged(newWorkspace);
+    });
+
+    // Dispose of the workspace when our div ref goes away (Equivalent to didComponentUnmount)
+    return () => {
+      newWorkspace.dispose();
+    };
+  }, [handleWorkspaceChanged, ref]);
+
+  // xmlDidChange callback
+  React.useEffect(() => {
+    if (workspace == null) {
+      return undefined;
+    }
+
+    const [callback, cancel] = debounce(() => {
+      const newXml = Blockly.Xml.domToText(
+        Blockly.Xml.workspaceToDom(workspace)
+      );
+      if (newXml === xml) {
+        return;
+      }
+
+      setXml(newXml);
+    }, 200);
+
+    workspace.addChangeListener(callback);
+
+    return () => {
+      workspace.removeChangeListener(callback);
+      cancel();
+    };
+  }, [workspace, xml]);
+
+  // Initial Xml Changes
+  React.useEffect(() => {
+    if (xml && workspace && !didInitialImport) {
+      const success = importFromXml(xml, workspace, onImportXmlError);
+      if (!success) {
+        setXml(null);
+      }
+      setDidInitialImport(true);
+    }
+  }, [xml, workspace, didInitialImport, onImportXmlError]);
+};
+
+export default useBlocklyWorkspace;
